pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "apbs_parser"
description = "NOMAD parser for APBS"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Fabian Li", email = "udskm@student.kit.edu" },
]
maintainers = [
    { name = "Fabian Li", email = "udskm@student.kit.edu" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab[parsing, infrastructure]>=1.3.10",
    "python-magic-bin; sys_platform == 'win32'",
    "nomad-schema-plugin-run>=1.0.1",
    "nomad-simulations>0.0.7",
]

[project.urls]
Repository = "https://github.com/fabianli789/apbs_parser"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint 
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
parser_entry_point = "apbs_parser.parsers:parser_entry_point"
schema_package_entry_point = "apbs_parser.schema_packages:schema_package_entry_point"
normalizer_entry_point = "apbs_parser.normalizers:normalizer_entry_point"
app_entry_point = "apbs_parser.apps:app_entry_point"
example_upload_entry_point = "apbs_parser.example_uploads:example_upload_entry_point"
[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
